+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [1.01] - 11/20/2022
 
-- Add devotionals
+- Added devotionals
 
 ## [1.00] - 11/18/2022