+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## [1.01] - 11/20/2022
+## [1.01] - 11/21/2022
 
 - Added devotionals
+- Added copy verse link
 - Added verse comparison
+- Added cross references
 
 ## [1.00] - 11/18/2022